[REFACTOR] items and pipelines packages
